Refactor cdn and fixes to arkadius comments
