+++ engine/demo/src/main/java/pl/touk/nussknacker/engine/javademo/DemoProcessConfigCreator.java
@@ -15,8 +15,8 @@ import pl.touk.nussknacker.engine.api.test.TestParsingUtils;
 import pl.touk.nussknacker.engine.demo.LoggingExceptionHandlerFactory;
 import pl.touk.nussknacker.engine.javaapi.process.ExpressionConfig;
 import pl.touk.nussknacker.engine.javaapi.process.ProcessConfigCreator;
-import pl.touk.nussknacker.engine.kafka.KafkaSinkFactory;
-import pl.touk.nussknacker.engine.kafka.KafkaSourceFactory;
+import pl.touk.nussknacker.engine.kafka.sink.KafkaSinkFactory;
+import pl.touk.nussknacker.engine.kafka.source.KafkaSourceFactory;
 import pl.touk.nussknacker.engine.kafka.serialization.KafkaSerializationSchemaFactory;
 import pl.touk.nussknacker.engine.kafka.serialization.schemas;
 import scala.Option;
