+++ engine/flink/avro-util/src/main/scala/pl/touk/nussknacker/engine/avro/schemaregistry/confluent/serialization/ConfluentKafkaAvroDeserializationSchemaFactory.scala
@@ -5,7 +5,7 @@ import org.apache.flink.api.common.typeinfo.TypeInformation
 import org.apache.kafka.common.serialization.Deserializer
 import pl.touk.nussknacker.engine.avro.schemaregistry.confluent.client.ConfluentSchemaRegistryClientFactory
 import pl.touk.nussknacker.engine.kafka.KafkaConfig
-import pl.touk.nussknacker.engine.kafka.serialization.{BaseKafkaDeserializationSchemaVersionAwareFactory, BaseKeyValueKafkaDeserializationSchemaVersionAwareFactory}
+import pl.touk.nussknacker.engine.kafka.serialization.{BaseKafkaVersionAwareDeserializationSchemaFactory, BaseKafkaVersionAwareKeyValueDeserializationSchemaFactory}
 
 trait ConfluentKafkaAvroDeserializerFactory {
   import collection.JavaConverters._
@@ -22,14 +22,14 @@ trait ConfluentKafkaAvroDeserializerFactory {
 }
 
 class ConfluentKafkaAvroDeserializationSchemaFactory[T: TypeInformation](schemaRegistryClientFactory: ConfluentSchemaRegistryClientFactory, useSpecificAvroReader: Boolean)
-  extends BaseKafkaDeserializationSchemaVersionAwareFactory[T] with ConfluentKafkaAvroDeserializerFactory {
+  extends BaseKafkaVersionAwareDeserializationSchemaFactory[T] with ConfluentKafkaAvroDeserializerFactory {
 
   override protected def createValueDeserializer(topics: List[String], version: Option[Int], kafkaConfig: KafkaConfig): Deserializer[T] =
     createDeserializer[T](schemaRegistryClientFactory, kafkaConfig, isKey = false, useSpecificAvroReader)
 }
 
 abstract class ConfluentKeyValueKafkaAvroDeserializationFactory[T: TypeInformation](schemaRegistryClientFactory: ConfluentSchemaRegistryClientFactory, useSpecificAvroReader: Boolean)
-  extends BaseKeyValueKafkaDeserializationSchemaVersionAwareFactory[T] with ConfluentKafkaAvroDeserializerFactory {
+  extends BaseKafkaVersionAwareKeyValueDeserializationSchemaFactory[T] with ConfluentKafkaAvroDeserializerFactory {
 
   override protected def createKeyDeserializer(topics: List[String], version: Option[Int], kafkaConfig: KafkaConfig): Deserializer[K] =
     createDeserializer[K](schemaRegistryClientFactory, kafkaConfig, isKey = true, useSpecificAvroReader)
