+++ engine/flink/kafka-util/src/main/scala/pl/touk/nussknacker/engine/kafka/KafkaFactory.scala
@@ -1,0 +1,5 @@
+package pl.touk.nussknacker.engine.kafka
+
+object KafkaFactory {
+  final val TopicParamName = "topic"
+}
