+++ engine/flink/kafka-util/src/test/scala/pl/touk/nussknacker/engine/kafka/KafkaSourceFactorySpec.scala
@@ -10,6 +10,7 @@ import pl.touk.nussknacker.engine.api.namespaces.DefaultObjectNaming
 import pl.touk.nussknacker.engine.api.process.ProcessObjectDependencies
 import pl.touk.nussknacker.engine.api.test.TestParsingUtils
 import pl.touk.nussknacker.engine.api.{MetaData, StreamMetaData}
+import pl.touk.nussknacker.engine.kafka.source.KafkaSourceFactory
 
 class KafkaSourceFactorySpec extends FlatSpec with BeforeAndAfterAll with KafkaSpec with Matchers {
